+++ frontend/src/_test_/SideBar.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import SideBar from '../../Components/SideBar';
+import SideBar from '../Components/SideBar';
 
 configure({ adapter: new Adapter() });
