+++ frontend/src/_test_/Spinner.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import CircularIndeterminate from '../../Components/Spinner';
+import CircularIndeterminate from '../Components/Spinner';
 
 configure({ adapter: new Adapter() });
