+++ frontend/src/_test_/TopNav.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import TopNav from '../../Components/TopNav';
+import TopNav from '../Components/TopNav';
 
 configure({ adapter: new Adapter() });
